tests/Interop/Windows/um/wincodec/IWICColorContextTests.cs
// Copyright © Tanner Gooding and Contributors. Licensed under the MIT License (MIT). See License.md in the repository root for more information.

using System;
using System.Runtime.InteropServices;
using NUnit.Framework;
using static TerraFX.Interop.WinCodec;

namespace TerraFX.Interop.UnitTests
{
    /// <summary>Provides validation of the <see cref="IWICColorContext" /> struct.</summary>
    public static class IWICColorContextTests
    {
        /// <summary>Validates that the <see cref="Guid" /> of the <see cref="IWICColorContext" /> struct is correct.</summary>
        [Test]
        public static void GuidOfTest()
        {
            Assert.That(typeof(IWICColorContext).GUID, Is.EqualTo(IID_IWICColorContext));
        }

        /// <summary>Validates that the layout of the <see cref="IWICColorContext" /> struct is <see cref="LayoutKind.Sequential" />.</summary>
        [Test]
        public static void IsLayoutSequentialTest()
        {
            Assert.That(typeof(IWICColorContext).IsLayoutSequential, Is.True);
        }

        /// <summary>Validates that the size of the <see cref="IWICColorContext" /> struct is correct.</summary>
        [Test]
        public static void SizeOfTest()
        {
            if (Environment.Is64BitProcess)
            {
                Assert.That(Marshal.SizeOf<IWICColorContext>(), Is.EqualTo(8));
            }
            else
            {
                Assert.That(Marshal.SizeOf<IWICColorContext>(), Is.EqualTo(4));
            }
        }

        /// <summary>Provides validation of the <see cref="IWICColorContext.Vtbl" /> struct.</summary>
        public static class VtblTests
        {
            /// <summary>Validates that the layout of the <see cref="IWICColorContext" /> struct is <see cref="LayoutKind.Sequential" />.</summary>
            [Test]
            public static void IsLayoutSequentialTest()
            {
                Assert.That(typeof(IWICColorContext.Vtbl).IsLayoutSequential, Is.True);
            }

            /// <summary>Validates that the size of the <see cref="IWICColorContext" /> struct is correct.</summary>
            [Test]
            public static void SizeOfTest()
            {
                if (Environment.Is64BitProcess)
                {
                    Assert.That(Marshal.SizeOf<IWICColorContext.Vtbl>(), Is.EqualTo(72));
                }
                else
                {
                    Assert.That(Marshal.SizeOf<IWICColorContext.Vtbl>(), Is.EqualTo(36));
                }
            }
        }
    }
}
